# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "flutter-agent",
    "summary": "Local Flutter requirement refinement service: turns one-liners into spec / architecture / task breakdowns using DeepSeek (or any OpenAI-compatible model) plus Markdown skill files.",
    "description": "Two ways to call:\n\n* **Canonical**: `POST /v1/refine` returns a strongly-typed `RefineResponse` (classify / spec / architecture / breakdown / acceptance + final markdown PRD).\n* **OpenAI-compatible**: `POST /v1/chat/completions` accepts the standard OpenAI chat format. Use `model=\"flutter-agent\"` to run the refinement pipeline, or any other model name to passthrough to the upstream provider.",
    "contact": {
      "name": "flutter-agent",
      "url": "http://127.0.0.1:8765/docs"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/skills": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "List all loaded skills",
        "operationId": "list_skills_v1_skills_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SkillSummary"
                  },
                  "title": "Response List Skills V1 Skills Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/skills/{skill_id}": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "Get a single skill (incl. full markdown body)",
        "operationId": "get_skill_v1_skills__skill_id__get",
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/skills/reload": {
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "Hot-reload skill markdown files from disk",
        "operationId": "reload_skills_v1_skills_reload_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Reload Skills V1 Skills Reload Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/skills/rank": {
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "Dry-run the skill selection for a requirement (no model calls)",
        "description": "Explain which skills would be injected for a requirement and why.\n\nRuns the exact same ranker the pipeline uses (keyword scores, foundational\npinning, family de-dup, token budget) and returns every skill's score and\nselection outcome — purely local, zero upstream tokens.",
        "operationId": "rank_skills_for_requirement_v1_skills_rank_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillRankRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillRankResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/refine": {
      "post": {
        "tags": [
          "refine"
        ],
        "summary": "Refine a Flutter requirement into spec / architecture / tasks",
        "description": "Run the multi-stage refinement pipeline against the configured upstream.",
        "operationId": "refine_v1_refine_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefineResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream model error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/refine/stream": {
      "post": {
        "tags": [
          "refine"
        ],
        "summary": "Refine with live SSE progress (stage-by-stage events + final result)",
        "description": "Run the pipeline while streaming progress events over SSE.\n\nStages can take tens of seconds each against a real upstream model; this\nendpoint lets clients render live progress instead of a blank wait.",
        "operationId": "refine_stream_v1_refine_stream_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "text/event-stream of JSON events: pipeline_start, stage_start, stage_complete (elapsed/usage/cost), cache_hit, done (full RefineResponse), error. Terminated by `data: [DONE]`.",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {}
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Platform": {
        "type": "string",
        "enum": [
          "mobile",
          "desktop",
          "web",
          "auto"
        ],
        "title": "Platform"
      },
      "RefineRequest": {
        "properties": {
          "requirement": {
            "type": "string",
            "minLength": 2,
            "title": "Requirement",
            "description": "The raw user requirement (one liner or paragraph)."
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/Platform"
            },
            "type": "array",
            "title": "Platforms",
            "description": "Target platforms. Use ['auto'] to let the model decide."
          },
          "skills": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skills",
            "description": "Override skill selection. If null, the classify stage decides. Pass skill ids, e.g. ['flutter-mobile','task-refinement']."
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/Stage"
            },
            "type": "array",
            "title": "Stages",
            "description": "Which pipeline stages to run, in order."
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature",
            "description": "Override sampling temperature."
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 32000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens",
            "description": "Override per-call max tokens."
          },
          "extra_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Context",
            "description": "Free-form extra context to append after the user requirement."
          },
          "use_cache": {
            "type": "boolean",
            "title": "Use Cache",
            "description": "If true, reuse the most recent run with an identical input (requirement+skills+stages+temperature+max_tokens+model). Cache hits are returned with cached=true and skip the upstream call.",
            "default": false
          },
          "validate_packages": {
            "type": "boolean",
            "title": "Validate Packages",
            "description": "If true, every entry in architecture.third_party is checked against pub.dev to catch hallucinated package names.",
            "default": true
          },
          "review_max_iterations": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Review Max Iterations",
            "description": "Closed loop: when the review stage flags blocking findings (severity blocker/major or blocking=true), re-run implementation with the findings as feedback and re-review, up to this many extra passes. 0 disables the loop (review stays advisory).",
            "default": 1
          },
          "review_block_severity": {
            "type": "string",
            "enum": [
              "blocker",
              "major",
              "minor"
            ],
            "title": "Review Block Severity",
            "description": "Lowest finding severity that makes the review blocking. 'major' (default) blocks on blocker+major; 'blocker' only on blockers (loosest); 'minor' also blocks on minor (strictest). The model's explicit blocking=true always blocks regardless of this setting.",
            "default": "major"
          }
        },
        "type": "object",
        "required": [
          "requirement"
        ],
        "title": "RefineRequest",
        "description": "User-facing request for the /v1/refine endpoint."
      },
...
      "SkillDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique skill id, matches directory name."
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": "1.0.0"
          },
          "platforms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Platforms"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "applies_when": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applies When"
          },
          "stage_hints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Stage Hints"
          },
          "extends": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Extends",
            "description": "Skill ids this skill deepens/specializes. The ranker treats self + targets as one family and de-duplicates so a broad parent and its narrow child do not both consume the token budget."
          },
          "see_also": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "See Also",
            "description": "Related skill ids (cross-references; non-binding for ranking)."
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "body": {
            "type": "string",
            "title": "Body",
            "description": "Markdown body (everything after front-matter)."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "path",
          "body"
        ],
        "title": "SkillDetail"
      },
      "SkillRankItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "score": {
            "type": "number",
            "title": "Score",
            "description": "Relevance score from the keyword ranker."
          },
          "selected": {
            "type": "boolean",
            "title": "Selected",
            "description": "Whether select_within_budget kept it."
          },
          "foundational": {
            "type": "boolean",
            "title": "Foundational",
            "description": "Whether the skill was force-included as foundational."
          },
          "family": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family",
            "description": "Family root when the skill belongs to an `extends` family."
          },
          "estimated_tokens": {
            "type": "integer",
            "title": "Estimated Tokens",
            "description": "Rough token estimate of the skill body."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "score",
          "selected",
          "foundational",
          "estimated_tokens"
        ],
        "title": "SkillRankItem",
        "description": "One skill's standing in the ranking."
      },
      "SkillRankRequest": {
        "properties": {
          "requirement": {
            "type": "string",
            "minLength": 2,
            "title": "Requirement",
            "description": "The raw user requirement to rank skills against."
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/Platform"
            },
            "type": "array",
            "title": "Platforms",
            "description": "Target platforms. Use ['auto'] for the default set."
          },
          "token_budget": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Budget",
            "description": "Override the skill token budget (default 40000)."
          }
        },
        "type": "object",
        "required": [
          "requirement"
        ],
        "title": "SkillRankRequest",
        "description": "Dry-run the skill selection for a requirement — no model calls."
      },
      "SkillRankResponse": {
        "properties": {
          "requirement": {
            "type": "string",
            "title": "Requirement"
          },
          "platforms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Platforms"
          },
          "token_budget": {
            "type": "integer",
            "title": "Token Budget"
          },
          "foundational": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Foundational",
            "description": "Skill ids force-included for this requirement."
          },
          "selected": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selected",
            "description": "Skill ids that survive ranking + budget trimming, in order."
          },
          "ranked": {
            "items": {
              "$ref": "#/components/schemas/SkillRankItem"
            },
            "type": "array",
            "title": "Ranked",
            "description": "All skills with score / selection / family details."
          }
        },
        "type": "object",
        "required": [
          "requirement",
          "platforms",
          "token_budget",
          "foundational",
          "selected",
          "ranked"
        ],
        "title": "SkillRankResponse",
        "description": "Explainable result of the skill selection dry-run."
      },
      "SkillSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique skill id, matches directory name."
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": "1.0.0"
          },
          "platforms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Platforms"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "applies_when": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applies When"
          },
          "stage_hints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Stage Hints"
          },
          "extends": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Extends",
            "description": "Skill ids this skill deepens/specializes. The ranker treats self + targets as one family and de-duplicates so a broad parent and its narrow child do not both consume the token budget."
          },
          "see_also": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "See Also",
            "description": "Related skill ids (cross-references; non-binding for ranking)."
          },
          "path": {
            "type": "string",
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "path"
        ],
        "title": "SkillSummary",
        "description": "Skill metadata + path on disk; body is omitted to keep listings light."
      },
      "Stage": {
        "type": "string",
        "enum": [
          "classify",
          "spec",
          "architecture",
          "breakdown",
          "implementation",
          "review",
          "acceptance",
          "markdown"
        ],
        "title": "Stage"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}